+++ internal/collector/collector.go
@@ -89,6 +89,7 @@ func (c *Collector) RegisterServiceCollectors() {
 		c.cfg.Collectors.AP.Traffic,
 		c.cfg.Collectors.AP.Errors,
 		c.cfg.Collectors.AP.Join,
+		c.cfg.Collectors.AP.Spectrum,
 		c.cfg.Collectors.AP.Info,
 	) {
 		apSource := wnc.NewAPSource(c.sharedDataSource)
@@ -165,6 +166,7 @@ func (c *Collector) registerAPCollector(apSource wnc.APSource, rrmSource wnc.RRM
 		Traffic:    c.cfg.Collectors.AP.Traffic,
 		Errors:     c.cfg.Collectors.AP.Errors,
 		Join:       c.cfg.Collectors.AP.Join,
+		Spectrum:   c.cfg.Collectors.AP.Spectrum,
 		Info:       c.cfg.Collectors.AP.Info,
 		InfoLabels: c.cfg.Collectors.AP.InfoLabels,
 	})